requirements.txt
# requirements.txt

streamlit==1.43.0
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
plotly==6.0.0
openpyxl==3.1.5
xlsxwriter==3.2.2
numpy==2.2.3
reportlab==4.1.0
PyPDF2==3.0.1
requests==2.32.3
pymongo==4.11.1
qrcode==8.0
pillow==10.0.0
